Split Evaluation.kt into multiple files
